Move formatting rules into a separate file
