+++ src/python/pants/backend/python/lint/ruff/rules.py
@@ -59,41 +59,11 @@ class RuffFixRequest(FixTargetsRequest):
         return RuffLintRequest.tool_id
 
 
-class RuffFormatRequest(FmtTargetsRequest):
-    field_set_type = RuffFieldSet
-    tool_subsystem = Ruff
-    partitioner_type = PartitionerType.DEFAULT_SINGLE_PARTITION
-
-    @classproperty
-    def tool_name(cls) -> str:
-        return "ruff format"
-
-    @classproperty
-    def tool_id(self) -> str:
-        return "ruff-format"
-
-
 @dataclass(frozen=True)
 class _RunRuffRequest:
     snapshot: Snapshot
     mode: RuffMode
     interpreter_constraints: Optional[InterpreterConstraints] = None
-
-
-# Note - this function is kept separate because it is invoked from update_build_files.py, but
-# not as a rule.
-async def _run_ruff_fmt(
-    request: AbstractFmtRequest.Batch,
-    ruff: Ruff,
-    interpreter_constraints: Optional[InterpreterConstraints] = None,
-) -> FmtResult:
-    run_ruff_request = _RunRuffRequest(
-        snapshot=request.snapshot,
-        mode=RuffMode.FORMAT,
-        interpreter_constraints=interpreter_constraints,
-    )
-    result = await _run_ruff(run_ruff_request, ruff)
-    return await FmtResult.create(request, result)
 
 
 async def _run_ruff(
@@ -169,16 +139,10 @@ async def ruff_lint(request: RuffLintRequest.Batch[RuffFieldSet, Any], ruff: Ruf
     return LintResult.create(request, result)
 
 
-@rule(desc="Format with `ruff format`", level=LogLevel.DEBUG)
-async def ruff_fmt(request: RuffFormatRequest.Batch, ruff: Ruff) -> FmtResult:
-    return await _run_ruff_fmt(request, ruff)
-
-
 def rules():
     return [
         *collect_rules(),
         *RuffFixRequest.rules(),
         *RuffLintRequest.rules(),
-        *RuffFormatRequest.rules(),
         *pex.rules(),
     ]
